requirements.txt
streamlit>=1.28.0
docxtpl>=0.16.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
Pillow>=10.0.0
jinja2>=3.1.0

# IA
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
